Moved isOrderUnfillable to order utils
